create a derived class AddEmployeePanel
